Gave up
Gave up on the project.

I am soo done.
Look there's an issue. When I click delete on a task that's not at the end, the state and local storage updates. However, the UI doesn't UPDATE! The task that's clicked isn't deleted! When I refresh the page, the task are correctly displayed.

I give up. I probably don't know enough of react for this project.