PROJECT DONE
Project is done. I finished the project. I learned a valuable lesson. Never use the index of an array to set the key prop or id for a component.

I had an issue because I was using the index of the array as the key prop and id. To fix the error, I just download a npm package that generated ids. The issue is now fixed.